port the core plugins that use a Neigborhood3x3Operation to be compatible with imglib2
git-svn-id: http://code.imagej.net/svn/imagej/branches/imglib2@2714 be23795f-a60e-40d4-ba19-055798872e57

+++ core/core-plugins/src/main/java/imagej/core/plugins/neigh/FindEdges.java
@@ -81,7 +81,7 @@ public class FindEdges implements ImageJPlugin {
 
 		/** at each new neighborhood start tracking neighbor 0 */
 		@Override
-		public void initializeNeighborhood(int[] position) {
+		public void initializeNeighborhood(long[] position) {
 			// nothing to do
 		}
